pyproject.toml
[project]
dependencies = [
    "fastapi[standard]>=0.135.2",
    "python-datauri>=3.0.2",
    "openai>=2.29.0",

    "librosa>=0.11.0",
    "transformers>=4.36",

    "torch==2.8.0",
    "torchvision==0.23.0",
    "torchaudio==2.8.0",
    "intel-extension-for-pytorch==2.8.10+xpu",
    "oneccl_bind_pt==2.8.0+xpu",

    "accelerate>=1.12.0",
]
name = "qwen3-voice-embedding-server"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "Wei Ji (FlySkyPie)", email = "c445dj544@gmail.com" }]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false

[tool.fastapi]
entrypoint = "src.main:app"

[tool.pdm.scripts]
start.cmd = "fastapi run"
start.env.OverrideDefaultFP64Settings = "1"
start.env.IGC_EnableDPEmulation = "1"

dev.cmd = "fastapi dev"
dev.env.OverrideDefaultFP64Settings = "1"
dev.env.IGC_EnableDPEmulation = "1"

[[tool.pdm.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/xpu"
include_packages = ["torch", "torchvision", "torchaudio", "pytorch-triton-xpu"]
exclude_packages = ["*"]

[[tool.pdm.source]]
name = "intel"
url = "https://pytorch-extension.intel.com/release-whl/stable/xpu/us/"
include_packages = ["intel-extension-for-pytorch", "oneccl_bind_pt"]
exclude_packages = ["*"]
